[WIP] Refactors + Tooling (#43)
The current testing is okay but I was noticing some deviations in the way it schedules transactions too far in the future and didn't know how to change that. So I started another tool to `spamTx` which is similar but makes more sense to me. This also led to some refactoring to make some sense out of the codebase. Since porting to TS might take too much time (as we've learned from eac.js-client), I'm just focusing on reorganizing the code rather than typing it.